feat/refactor: graceful shutdown
