+++ src/services/invoice-service.ts
@@ -1,4 +1,4 @@
-import { execute } from "../utils/mysql.connector";
+import { execute } from "../utils/database-connector";
 import { Invoice, InvoiceStatus } from "../classes/invoice";
 import { invoiceQueries } from "../queries/invoice-queries";
 import { setInterval } from "timers";
